% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{echo_vec}
\alias{echo_vec}
\title{Print an atomic vector as your input.}
\usage{
echo_vec(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\value{
An atomic vector as your input.
}
\description{
Print an atomic vector as your input.
}
\examples{
echo_vec(letters)
}
