% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_utils.R
\name{extract_gene_id_from_gff}
\alias{extract_gene_id_from_gff}
\title{Extract gene IDs from GTF/GFF file.}
\usage{
extract_gene_id_from_gff(gff_file, target_type = "gene")
}
\arguments{
\item{gff_file}{character, path to GTF/GFF file.}

\item{target_type}{character, which type (gene or transcript) to be extracted.}
}
\value{
A data frame.
}
\description{
Extract gene IDs from GTF/GFF file.
}
