% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{file_transfer}
\alias{file_transfer}
\title{Transfer files in various modes.}
\usage{
file_transfer(from, to, ops = c("soft", "hard", "copy", "move"))
}
\arguments{
\item{from}{character vector, containing file names.}

\item{to}{same as \code{from}.}

\item{ops}{character vector, containing operations. If the
previous one cannot be run successfully over all files,
then the latter will be run until some operation can be run
successfully, or it will raise an error.}
}
\value{
A list containing \code{from} and \code{to}.
}
\description{
Transfer files in various modes.
}
\examples{
\dontrun{
from <- c("file1.txt", "file2.txt", "file3.txt")
to <- c("temp1.txt", "temp2.txt", "temp3.txt")
file_transfer(from, to)
}
}
