% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{group_enriched_terms}
\alias{group_enriched_terms}
\title{Group GO terms via hierarchical clustering}
\usage{
group_enriched_terms(
  df,
  out_file_prefix = "",
  x_column = "GeneRatio",
  fill_column = "p.adjust",
  size_column = "Count",
  heatmap_data = c("sim", "dissim"),
  heatmap_width = 500,
  heatmap_height = 500,
  cut_hc_height = 0.25
)
}
\arguments{
\item{df}{a data frame. At present, we assume that you should get it from the enriched results of clusterProfiler.}

\item{out_file_prefix}{a character string. This should contain both the output directory and the output file prefix.}

\item{x_column}{which column in \code{df} is used as the x-axis of the dot plot.}

\item{fill_column}{which column in \code{df} is used as the \code{fill} aesthetics in dot plot.}

\item{size_column}{which column in \code{df} is used as the \code{size} aesthetics in dot plot.}

\item{heatmap_data}{which type of data should be used to plot heatmap. Can be either
\code{sim} (using similarity matrix) or \code{dissim} (using dissimilarity matrix).}

\item{heatmap_width}{the maximum width of heatmap in inches.}

\item{heatmap_height}{the maximum height of heatmap in inches.}

\item{cut_hc_height}{the height threshold for cutting the dendrogram.}
}
\value{
A list containing all data/plot objects used/generated.
}
\description{
This function first calculates a dissimilarity matrix via \code{cluster::daisy} from an asymmetric binary matrix,
in which rows are terms and columns are genes and each value can be either
1 (the term contains this gene) or 0 (vice versa). Then this dissimilarity matrix
is used to perform hierarchical clustering using \code{cluster::agnes}. Finally,
we cut the dendrogram based on a given height threshold into modules. In each module,
terms have a significantly high number of common genes, compared to other terms, not in the same module.
This means that terms in the same module may be redundant and you can only pick some of them
based on your scientific question for further analysis.
}
