% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{ppreview}
\alias{ppreview}
\title{Preview an R plot and resize it before saving it interactively}
\usage{
ppreview(plot, file, width = 1200, height = 1200)
}
\arguments{
\item{plot}{a plot object.}

\item{file}{file path to save plot.}

\item{width}{the width of the plot in pixels (1200 by default).}

\item{height}{the height of the plot in pixels (1200 by default).}
}
\value{
NULL.
}
\description{
The plot object is previewed as a PNG format, and then
if the plot size is nice, it will be saved in the given format,
inferred from the file extension.
}
\examples{
\dontrun{
library(ggplot2)

df <- data.frame(
    x = rnorm(100),
    y = rnorm(100)
)

p <- ggplot(df, aes(x, y)) +
    geom_point()

ppreview(p, "test.pdf")
}
}
