% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{set_fonts}
\alias{set_fonts}
\title{Provide an easy-to-use interface to add fonts}
\usage{
set_fonts(
  font_families = c("Arial", "Times New Roman"),
  google_fonts = NULL,
  load_only = FALSE,
  ...
)
}
\arguments{
\item{font_families}{a character vector or a data frame. If it is a character vector,
then the font files will be retrieved from \code{sysfonts::font_files()}; otherwise the
given data frame must have exactly the same forms as retrieved from \code{sysfonts::font_files()},
except that the \code{file} column must be absolute paths.}

\item{google_fonts}{a character vector. If a named vector is given, the names will be used as family names.}

\item{load_only}{logical. If \code{TRUE}, call \code{showtext_auto}.}

\item{...}{parameters passed to \code{font_add_google()}.}
}
\value{
NULL.
}
\description{
You can add fonts from either \code{sysfonts::font_files()} or local font files
by giving \code{font_families}, and/or add google fonts by giving \code{google_fonts}
(see \href{https://fonts.google.com}{Google fonts}).
}
\examples{
\dontrun{
set_fonts(font_families = c("Arial", "Times New Roman"))
set_fonts(google_fonts = c("bell" = "Schoolbell"))
}
}
\seealso{
For more info, refer to the \code{showtext} and \code{sysfonts} packages.
}
